fix: __PATH_PREFIX__ not replaced when using user directive
Files were owned by node:node with 755. If you override the user then they won't have permission to replace __PATH_PREFIX__ on startup. This change makes the UI folder writable by anybody. At first I didn't like this, but as it was 755 before then the node user could already write files. Really this is no worse as we're in a container.