feat: Replace plex-api
plex-api used the request package under the hood which was deprecated
5 years ago and has numerous open issues around connection issues.
I took the code we were using from plex-api and replaced request with
the native fetch implementation.

I have also disabled keep alives as this resolved a socket hang up
issue for a user and was the main reason for looking at the plex-api
library. There are numerous reported issues with node and keep
alives resulting in socket hang ups and this was the easiest solution.
Given most people will be communicating with Plex on LAN this should not
be too much of an issue. Besides, all other APIs we talk to use axios
which does not have keep alives enabled out of the box.