feat: Add Jellyseerr support (#1606)
This is a complete copy of our existing Overseerr code as the APIs we use are functionally the same at the moment. The only feature I have not brought across is the 'Force reset Overseerr record' option, as availability sync should be used instead. If we have problems with this we can always introduce the option again.