fix: Sonarr returns monitored=true if non-existent
If the /series?tvdbId=$ API call returns nothing we would use a different endpoint as a fallback. This endpoint returns monitored: true as this is the default for a new series: https://github.com/Sonarr/Sonarr/issues/5761#issuecomment-1607959602. This issue becomes apparent when you have two Sonarr instances feeding a single Plex library and try to use the monitored condition.