fix: Do not bypass rules that error out
When getters return undefined, or rule comparison fails, we would skip over the rule completely, essentially returning a true result. This could result in media being incorrectly actioned, should the failure persist for the configured take action after days. It also means that temporary issues, such as an API being unavaliable, could cause an add or remove action because of the logic change. This action would subsequently be reversed on the next rule handler execution. With the recent addition of API retries this should hopefully not be much of an issue, though we still need to add retries to the Plex API.

With this change, an error occuring in a rule for a media item will stop further rule evaulation for it, treating it as a 'false' result.